// packages/web/src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      admins: {
        Row: {
          created_at: string;
          user_id: string;
        };
        Insert: {
          created_at?: string;
          user_id: string;
        };
        Update: {
          created_at?: string;
          user_id?: string;
        };
        Relationships: [];
      };
      authors: {
        Row: {
          created_at: string;
          description: string | null;
          id: number;
          mindify_ai: boolean;
          name: string;
          production: boolean;
          slug: string;
        };
        Insert: {
          created_at?: string;
          description?: string | null;
          id?: number;
          mindify_ai?: boolean;
          name: string;
          production?: boolean;
          slug: string;
        };
        Update: {
          created_at?: string;
          description?: string | null;
          id?: number;
          mindify_ai?: boolean;
          name?: string;
          production?: boolean;
          slug?: string;
        };
        Relationships: [];
      };
// ...
      summaries: {
        Row: {
          author_id: number;
          chapters_id: number | null;
          created_at: string;
          id: number;
          image_url: string | null;
          mindify_ai: boolean | null;
          production: boolean;
          reading_time: number | null;
          slug: string;
          source_type: Database["public"]["Enums"]["source"];
          source_url: string | null;
          title: string;
          topic_id: number;
        };
        Insert: {
          author_id: number;
          chapters_id?: number | null;
          created_at?: string;
          id?: number;
          image_url?: string | null;
          mindify_ai?: boolean | null;
          production?: boolean;
          reading_time?: number | null;
          slug: string;
          source_type: Database["public"]["Enums"]["source"];
          source_url?: string | null;
          title: string;
          topic_id: number;
        };
        Update: {
          author_id?: number;
          chapters_id?: number | null;
          created_at?: string;
          id?: number;
          image_url?: string | null;
          mindify_ai?: boolean | null;
          production?: boolean;
          reading_time?: number | null;
          slug?: string;
          source_type?: Database["public"]["Enums"]["source"];
          source_url?: string | null;
          title?: string;
          topic_id?: number;
        };
        Relationships: [
          {
            foreignKeyName: "summaries_author_id_fkey";
            columns: ["author_id"];
            isOneToOne: false;
            referencedRelation: "authors";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "summaries_chapters_id_fkey";
            columns: ["chapters_id"];
            isOneToOne: false;
            referencedRelation: "chapters";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "summaries_topic_id_fkey";
            columns: ["topic_id"];
            isOneToOne: false;
            referencedRelation: "topics";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
      summary_requests: {
        Row: {
          author: string;
          created_at: string | null;
          id: number;
          source: Database["public"]["Enums"]["source"] | null;
          title: string;
          topic_id: number | null;
          user_id: string | null;
          validated: boolean;
        };
// ...
      };
// ...
      topics: {
        Row: {
          black_icon: string | null;
          created_at: string;
          emoji: string | null;
          id: number;
          name: string;
          production: boolean;
          slug: string;
          white_icon: string | null;
        };
        Insert: {
          black_icon?: string | null;
          created_at?: string;
          emoji?: string | null;
          id?: number;
          name: string;
          production?: boolean;
          slug: string;
          white_icon?: string | null;
        };
        Update: {
          black_icon?: string | null;
          created_at?: string;
          emoji?: string | null;
          id?: number;
          name?: string;
          production?: boolean;
          slug?: string;
          white_icon?: string | null;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
};

type PublicSchema = Database[Extract<keyof Database, "public">];

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    ? (PublicSchema["Tables"] & PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends keyof PublicSchema["Enums"] | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never;